requirements.txt
numpy==1.14.5
opencv-python==4.1.0.25
matplotlib==3.0.3
tensorflow-gpu==1.10.0
jupyter==1.0.0
Pillow==5.3.0
